+++ src/api/ADempiere/window.js
@@ -452,391 +452,3 @@ export function requestCreateChatEntry({
       return convertChatEntry(chatEntryResponse)
     })
 }
-
-/**
- * Issues exist
- * @param {string}  tableName
- * @param {number}  recordId
- * @param {string}  recordUuid
- */
-export function requestExistsIssues({
-  tableName,
-  recordId,
-  recordUuid
-}) {
-  return request({
-    url: '/user-interface/component/issue/exists-issues',
-    method: 'get',
-    params: {
-      record_id: recordId,
-      record_uuid: recordUuid,
-      table_name: tableName
-    }
-  })
-    .then(existsIssues => {
-      return existsIssues
-    })
-}
-
-/**
- * Issues List from Window
- * @param {string}  tableName
- * @param {number}  recordId
- * @param {string}  recordUuid
- */
-export function requestListExists({
-  tableName,
-  recordId,
-  recordUuid
-}) {
-  return request({
-    url: '/user-interface/component/issue/list-issues',
-    method: 'post',
-    data: {
-      record_id: recordId,
-      record_uuid: recordUuid,
-      table_name: tableName
-    }
-  })
-    .then(listExists => {
-      return listExists
-    })
-}
-
-/**
- * List Sales Representatives
- */
-export function listSalesRepresentatives() {
-  return request({
-    url: '/user-interface/component/issue/list-sales-representatives',
-    method: 'post',
-    data: {}
-  })
-    .then(salesRepresentatives => {
-      return salesRepresentatives
-    })
-}
-
-/**
- * List Request Types
- */
-export function listRequestTypes() {
-  return request({
-    url: '/user-interface/component/issue/list-request-types',
-    method: 'post',
-    data: {}
-  })
-    .then(requestTypes => {
-      return requestTypes
-    })
-}
-
-/**
- * List Statuses
- */
-export function listStatus({
-  requestTypeId
-}) {
-  return request({
-    url: '/user-interface/component/issue/list-statuses',
-    method: 'post',
-    data: {
-      request_type_id: requestTypeId
-    }
-  })
-    .then(listStatus => {
-      return listStatus
-    })
-}
-
-/**
- * List Priorities
- */
-export function listPriorities() {
-  return request({
-    url: '/user-interface/component/issue/list-priorities',
-    method: 'post',
-    data: {}
-  })
-    .then(listPriorities => {
-      return listPriorities
-    })
-}
-
-/**
- * New Issues
- * @param {string}  tableName
- * @param {number}  recordId
- * @param {string}  recordUuid
- */
-export function createIssues({
-  tableName,
-  recordId,
-  recordUuid,
-  subject,
-  summary,
-  requestTypeId,
-  requestTypeUuid,
-  salesRepresentativeId,
-  salesRepresentativeUuid,
-  statusId,
-  statusUuid,
-  priorityValue
-}) {
-  return request({
-    url: '/user-interface/component/issue/create-issue',
-    method: 'post',
-    data: {
-      record_id: recordId,
-      record_uuid: recordUuid,
-      table_name: tableName,
-      subject,
-      summary,
-      request_type_id: requestTypeId,
-      request_type_uuid: requestTypeUuid,
-      sales_representative_id: salesRepresentativeId,
-      sales_representative_uuid: salesRepresentativeUuid,
-      status_id: statusId,
-      status_uuid: statusUuid,
-      priority_value: priorityValue
-    }
-  })
-    .then(listExists => {
-      return listExists
-    })
-}
-
-/**
- * POST Update Issue
- *
- * req.query.token - user token
- * req.query.language - login language
- * req.body.id - id of record
- * req.body.uuid - uuid of record
- * req.body.subject - subject
- * req.body.summary - summary
- * req.body.requestTypeId - request type id
- * req.body.requestTypeUuid - request type uuid
- */
-export function updateIssues({
-  tableName,
-  id,
-  uuid,
-  subject,
-  summary,
-  requestTypeId,
-  requestTypeUuid,
-  salesRepresentativeId,
-  salesRepresentativeUuid,
-  statusId,
-  statusUuid,
-  priorityValue
-}) {
-  return request({
-    url: '/user-interface/component/issue/update-issue',
-    method: 'post',
-    data: {
-      id,
-      uuid,
-      table_name: tableName,
-      subject,
-      summary,
-      request_type_id: requestTypeId,
-      request_type_uuid: requestTypeUuid,
-      sales_representative_id: salesRepresentativeId,
-      sales_representative_uuid: salesRepresentativeUuid,
-      status_id: statusId,
-      status_uuid: statusUuid,
-      priority_value: priorityValue
-    }
-  })
-    .then(listExists => {
-      return listExists
-    })
-}
-
-/**
- * POST List Issue Comments
- *
- * req.query.token - user token
- * req.query.language - login language
- * req.body.issue_id - id of record parent issue
- * req.body.issue_uuid - uuid of record parent issued
- * req.query.page_size - size of page (customized)
- * req.query.page_token - token of page (optional for get a specific page)
- */
-export function listIssueComments({
-  issueId,
-  issueUuid
-}) {
-  return request({
-    url: '/user-interface/component/issue/list-issue-comments',
-    method: 'post',
-    data: {
-      issue_id: issueId,
-      issue_uuid: issueUuid
-    }
-  })
-    .then(listComments => {
-      return listComments
-    })
-}
-
-/**
- * POST Update Issue
- *
- * req.query.token - user token
- * req.query.language - login language
- * req.body.id - id of record
- * req.body.uuid - uuid of record
- * req.body.subject - subject
- * req.body.summary - summary
- * req.body.requestTypeId - request type id
- * req.body.requestTypeUuid - request type uuid
- */
-export function updateIssue({
-  id,
-  uuid,
-  subject,
-  summary,
-  requestTypeId,
-  requestTypeUuid,
-  salesRepresentativeId,
-  salesRepresentativeUuid,
-  statusId,
-  statusUuid,
-  priorityValue
-}) {
-  return request({
-    url: '/user-interface/component/issue/update-issue',
-    method: 'post',
-    data: {
-      id,
-      uuid,
-      subject,
-      summary,
-      request_type_id: requestTypeId,
-      request_type_uuid: requestTypeUuid,
-      sales_representative_id: salesRepresentativeId,
-      sales_representative_uuid: salesRepresentativeUuid,
-      status_id: statusId,
-      status_uuid: statusUuid,
-      priority_value: priorityValue
-    }
-  })
-    .then(listComments => {
-      return listComments
-    })
-}
-
-/**
- * POST Delete Issue
- *
- * req.query.token - user token
- * req.query.language - login language
- * req.body.id - id of record
- * req.body.uuid - uuid of record
- */
-export function deleteIssue({
-  id,
-  uuid
-}) {
-  return request({
-    url: '/user-interface/component/issue/delete-issue',
-    method: 'post',
-    data: {
-      id,
-      uuid
-    }
-  })
-    .then(listComments => {
-      return listComments
-    })
-}
-
-/**
- * POST Create Issue Comment
- *
- * req.query.token - user token
- * req.query.language - login language
- * req.body.issue_id - id of record parent issue
- * req.body.issue_uuid - uuid of record parent issued
- * req.body.result - result
- * req.body.dateNextAction - date to next action
- */
-export function createIssueComment({
-  issueId,
-  issueUuid,
-  result,
-  dateNextAction
-}) {
-  return request({
-    url: '/user-interface/component/issue/create-issue-comment',
-    method: 'post',
-    data: {
-      issue_id: issueId,
-      issue_uuid: issueUuid,
-      result,
-      date_next_action: dateNextAction
-    }
-  })
-    .then(listComments => {
-      return listComments
-    })
-}
-
-/**
-  * POST Update Issue Comment
-  *
-  * req.query.token - user token
-  * req.query.language - login language
-  * req.body.id - id of record
-  * req.body.uuid - uuid of record
-  * req.body.result - result
-  * req.body.dateNextAction - date to next action
- */
-export function updateIssueComment({
-  issueId,
-  issueUuid,
-  result,
-  dateNextAction
-}) {
-  return request({
-    url: '/user-interface/component/issue/update-issue-comment',
-    method: 'post',
-    data: {
-      id: issueId,
-      uuid: issueUuid,
-      result,
-      date_next_action: dateNextAction
-    }
-  })
-    .then(listComments => {
-      return listComments
-    })
-}
-
-/**
-  * POST Delete Issue Comments
-  *
-  * req.query.token - user token
-  * req.query.language - login language
-  * req.body.id - id of record
-  * req.body.uuid - uuid of record
- */
-export function deleteIssueComment({
-  issueId,
-  issueUuid
-}) {
-  return request({
-    url: '/user-interface/component/issue/delete-issue-comment',
-    method: 'post',
-    data: {
-      id: issueId,
-      uuid: issueUuid
-    }
-  })
-    .then(listComments => {
-      return listComments
-    })
-}
